+++ changelog.md
@@ -1,4 +1,5 @@
 ## 2.5.0
+- [FEA][IOS] Add a parameter to control the height of recipe card view. 
 - [FEA][IOS] Recipe carousel
 
 ## 2.4.0