+++ changelog.md
@@ -1,5 +1,9 @@
 ## 2.5.0
 - [FEA][COR] Prevent infinit loading on basket preview if miam is not well configure
+- [FIX][IOS] Add missing template for recipe card loading view
+- [FEA][IOS] Add a parameter to control the height of recipe card view. 
+- [FEA][IOS] Recipe carousel
+
 ## 2.4.0
 - [FEA][COR] carousel view model
 - [FEA][AND] recipe carousel component